@@ -7,7 +7,7 @@ readme = "README.md"
 packages = [{include = "app"}]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"
 flask = "3.0.2"
 pytest = "8.0.2"
 icalendar = "5.0.12"
@@ -17,6 +17,7 @@ mammoth = "^1.11.0"
 pydantic = "^2.0.0"
 git-filter-repo = "^2.47.0"
 python-dotenv = "^1.0.0"
+typer = {extras = ["all"], version = "^0.12.0"}
 
 [tool.poetry.scripts]
 calendar-sync = "cli:main"
