@@ -3,13 +3,15 @@
 from pathlib import Path
 
 from app.exceptions import IngestionError
+from app.ingestion.summary import build_ingestion_summary
 from app.models.calendar import Calendar
+from app.models.ingestion import IngestionResult
 
 
 class JSONReader:
     """Reader for JSON calendar files."""
 
-    def read(self, path: Path) -> Calendar:
+    def read(self, path: Path) -> IngestionResult:
         """Read calendar from JSON file."""
         try:
             with open(path, "r", encoding="utf-8") as f:
@@ -21,6 +23,9 @@ class JSONReader:
 
         try:
             # Use Pydantic's JSON deserialization
-            return Calendar.model_validate(data)
+            calendar = Calendar.model_validate(data)
+            return IngestionResult(
+                calendar=calendar, summary=build_ingestion_summary(calendar)
+            )
         except Exception as e:
             raise IngestionError(f"Failed to parse JSON calendar: {e}") from e
