@@ -1,23 +1,24 @@
 """CLI commands package."""
 
-from cli.commands.config import config_command
-from cli.commands.delete import delete_command
-from cli.commands.git_setup import git_setup_command
-from cli.commands.info import info_command
-from cli.commands.ls import ls_command
-from cli.commands.sync import sync_command
-from cli.commands.publish import publish_command
-from cli.commands.restore import restore_command
-from cli.commands.template import template_command
-
+from cli.commands.config import config
+from cli.commands.delete import delete
+from cli.commands.diff import diff
+from cli.commands.git_setup import git_setup
+from cli.commands.info import info
+from cli.commands.ls import ls
+from cli.commands.sync import sync
+from cli.commands.publish import publish
+from cli.commands.restore import restore
+from cli.commands.template import template
 __all__ = [
-    "config_command",
-    "delete_command",
-    "git_setup_command",
-    "info_command",
-    "ls_command",
-    "sync_command",
-    "publish_command",
-    "restore_command",
-    "template_command",
+    "config",
+    "delete",
+    "diff",
+    "git_setup",
+    "info",
+    "ls",
+    "sync",
+    "publish",
+    "restore",
+    "template",
 ]
