@@ -1,55 +1,58 @@
 """Restore calendar to specific git commit."""
 
 import logging
-import sys
 from datetime import timezone
 
-from app.config import CalendarConfig
-from app.storage.calendar_repository import CalendarRepository
-from app.storage.calendar_storage import CalendarStorage
-from app.storage.git_service import GitService
-from cli.setup import setup_reader_registry
+import typer
+from typing_extensions import Annotated
+
+from cli.commands.diff import display_diff
+from cli.context import get_context
 
 logger = logging.getLogger(__name__)
 
 
-def restore_command(name: str, commit: str, force: bool = False) -> None:
-    """
-    Restore calendar to specific git commit, version number, or relative command.
-    
-    Args:
-        name: Calendar name
-        commit: Git commit hash, version number (#3 or 3), or relative command (latest, previous)
-        force: If True, skip confirmation prompt
-    """
-    config = CalendarConfig.from_env()
-    storage = CalendarStorage(config)
-    reader_registry = setup_reader_registry()
-    git_service = GitService(
-        config.calendar_dir,
-        remote_url=config.calendar_git_remote_url,
-    )
-    repository = CalendarRepository(
-        config.calendar_dir, storage, git_service, reader_registry
-    )
+def restore(
+    name: Annotated[
+        str,
+        typer.Argument(help="Calendar name"),
+    ],
+    commit: Annotated[
+        str,
+        typer.Argument(
+            help="Git commit hash, version number (#3 or 3), or relative command (latest, previous)"
+        ),
+    ],
+    force: Annotated[
+        bool,
+        typer.Option("--force", "-f", help="Skip confirmation prompt"),
+    ] = False,
+) -> None:
+    """Restore calendar to specific git commit, version number, or relative command."""
+    ctx = get_context()
+    repository = ctx.repository
+    git_service = ctx.git_service
+    config = ctx.config
 
     # Check if calendar has any versions in git history (works for deleted calendars too)
     versions = repository.list_calendar_versions(name)
     if not versions:
         logger.error(f"Calendar '{name}' not found in git history")
-        sys.exit(1)
+        raise typer.Exit(1)
 
     # Parse commit input - could be version number, relative command, or commit hash
     target_commit = None
     commit_display = commit
-    
+
     # Check for version number (#3 or 3)
     if commit.startswith("#"):
         try:
             version_num = int(commit[1:])
             if version_num < 1 or version_num > len(versions):
-                logger.error(f"Version #{version_num} not found. Available versions: 1-{len(versions)}")
-                sys.exit(1)
+                logger.error(
+                    f"Version #{version_num} not found. Available versions: 1-{len(versions)}"
+                )
+                raise typer.Exit(1)
             target_commit = versions[version_num - 1][0]  # versions are 0-indexed
             commit_display = f"#{version_num} ({target_commit[:7]})"
         except ValueError:
@@ -58,8 +61,10 @@ def restore_command(name: str, commit: str, force: bool = False) -> None:
         try:
             version_num = int(commit)
             if version_num < 1 or version_num > len(versions):
-                logger.error(f"Version #{version_num} not found. Available versions: 1-{len(versions)}")
-                sys.exit(1)
+                logger.error(
+                    f"Version #{version_num} not found. Available versions: 1-{len(versions)}"
+                )
+                raise typer.Exit(1)
             target_commit = versions[version_num - 1][0]
             commit_display = f"#{version_num} ({target_commit[:7]})"
         except ValueError:
@@ -69,8 +74,10 @@ def restore_command(name: str, commit: str, force: bool = False) -> None:
         commit_display = f"latest ({target_commit[:7]})"
     elif commit.lower() == "previous":
         if len(versions) < 2:
-            logger.error(f"Only {len(versions)} version(s) available. Cannot restore to previous.")
-            sys.exit(1)
+            logger.error(
+                f"Only {len(versions)} version(s) available. Cannot restore to previous."
+            )
+            raise typer.Exit(1)
         target_commit = versions[1][0]  # Second most recent version
         commit_display = f"previous ({target_commit[:7]})"
     else:
@@ -89,29 +96,41 @@ def restore_command(name: str, commit: str, force: bool = False) -> None:
         if v_hash == target_commit:
             commit_info = (v_hash, v_date, v_message)
             break
-    
+
     if commit_info is None:
         logger.error(f"Commit '{commit}' not found for calendar '{name}'")
-        sys.exit(1)
+        raise typer.Exit(1)
 
     commit_hash, commit_date, commit_message = commit_info
 
+    # Format commit date
+    if commit_date.tzinfo is None:
+        commit_date = commit_date.replace(tzinfo=timezone.utc)
+    date_str = commit_date.strftime("%Y-%m-%d %H:%M")
+
+    # Load current calendar (before restore) for diff
+    current_result = repository.load_calendar(name, config.default_format)
+    current_calendar = current_result.calendar if current_result else None
+
+    # Load target calendar version for diff preview
+    target_result = repository.load_calendar_by_commit(
+        name, target_commit, config.default_format
+    )
+    target_calendar = target_result.calendar if target_result else None
+
     # Show confirmation prompt unless --force is set
     if not force:
-        # Format commit date
-        if commit_date.tzinfo is None:
-            commit_date = commit_date.replace(tzinfo=timezone.utc)
-        date_str = commit_date.strftime("%Y-%m-%d %H:%M")
-        
-        print(f"Will restore calendar '{name}' to:")
+        print(f"\nRestore calendar '{name}' to:")
         print(f"  Version: {commit_display}")
         print(f"  Date: {date_str}")
         print(f"  Message: {commit_message}")
+
+        # Show diff preview
+        display_diff(current_calendar, target_calendar, "current", "restored")
+
         print()
-        
-        response = input("Continue? [y/N]: ")
-        if response.lower() not in ['y', 'yes']:
-            print("Restore cancelled.")
+        if not typer.confirm("Continue?"):
+            typer.echo("Restore cancelled.")
             return
 
     # Get calendar directory (even if it doesn't exist yet)
@@ -121,9 +140,18 @@ def restore_command(name: str, commit: str, force: bool = False) -> None:
     calendar_dir.mkdir(parents=True, exist_ok=True)
 
     # Restore entire directory from git (includes calendar.ics, metadata.json, etc.)
-    git_service = repository.git_service
     if git_service.restore_directory_version(calendar_dir, target_commit):
-        print(f"Calendar '{name}' restored to {commit_display}")
+        # Get full path for clickable link
+        calendar_path = (calendar_dir / f"calendar.{config.default_format}").resolve()
+
+        # Show diff after restore when using --force
+        if force:
+            display_diff(current_calendar, target_calendar, "previous", "restored")
+
+        print(
+            f"\n{typer.style('✓', fg=typer.colors.GREEN, bold=True)} Calendar restored to {commit_display}"
+        )
+        print(f"  {calendar_path}")
     else:
         logger.error(f"Failed to restore calendar '{name}' to {commit_display}")
-        sys.exit(1)
+        raise typer.Exit(1)
