@@ -4,6 +4,7 @@ import json
 import os
 import shutil
 import subprocess
+import sys
 import tempfile
 from pathlib import Path
 from unittest.mock import MagicMock, patch
@@ -11,13 +12,13 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from app.config import CalendarConfig
-from cli.commands.git_setup import (
-    check_gh_cli_available,
-    create_repo_with_gh,
-    get_github_username_from_gh,
-    git_setup_command,
-    write_to_env_file,
-)
+from cli.context import CLIContext, set_context
+
+# Import the module properly using sys.modules
+# (not using `import cli.commands.git_setup as module` because Python
+# resolves that to the function named git_setup, not the module)
+from cli.commands import git_setup as git_setup_func
+git_setup_module = sys.modules['cli.commands.git_setup']
 
 
 @pytest.fixture
@@ -37,6 +38,14 @@ def temp_env_file(tmp_path):
         env_file.unlink()
 
 
+@pytest.fixture
+def mock_context():
+    """Set up mock CLI context."""
+    ctx = CLIContext()
+    set_context(ctx)
+    return ctx
+
+
 def test_check_gh_cli_available():
     """Test checking if GitHub CLI is available."""
     with patch("subprocess.run") as mock_run:
@@ -45,20 +54,20 @@ def test_check_gh_cli_available():
             MagicMock(returncode=0),  # gh --version
             MagicMock(returncode=0),  # gh auth status
         ]
-        assert check_gh_cli_available() is True
+        assert git_setup_module._check_gh_cli_available() is True
 
         # Test when gh is not installed
         mock_run.side_effect = [
             MagicMock(returncode=1),  # gh --version fails
         ]
-        assert check_gh_cli_available() is False
+        assert git_setup_module._check_gh_cli_available() is False
 
         # Test when gh is installed but not authenticated
         mock_run.side_effect = [
             MagicMock(returncode=0),  # gh --version
             MagicMock(returncode=1),  # gh auth status fails
         ]
-        assert check_gh_cli_available() is False
+        assert git_setup_module._check_gh_cli_available() is False
 
 
 def test_get_github_username_from_gh():
@@ -68,12 +77,12 @@ def test_get_github_username_from_gh():
         mock_run.return_value = MagicMock(
             returncode=0, stdout=json.dumps(mock_data)
         )
-        username = get_github_username_from_gh()
+        username = git_setup_module._get_github_username_from_gh()
         assert username == "testuser"
 
         # Test when gh command fails
         mock_run.return_value = MagicMock(returncode=1)
-        username = get_github_username_from_gh()
+        username = git_setup_module._get_github_username_from_gh()
         assert username is None
 
 
@@ -81,7 +90,7 @@ def test_create_repo_with_gh(tmp_path):
     """Test creating repo with GitHub CLI."""
     with patch("subprocess.run") as mock_run:
         mock_run.return_value = MagicMock(returncode=0)
-        result = create_repo_with_gh("testuser", "test-repo", tmp_path)
+        result = git_setup_module._create_repo_with_gh("testuser", "test-repo", tmp_path)
         assert result is True
         assert mock_run.called
         assert "gh" in mock_run.call_args[0][0]
@@ -90,7 +99,7 @@ def test_create_repo_with_gh(tmp_path):
 
         # Test failure
         mock_run.return_value = MagicMock(returncode=1)
-        result = create_repo_with_gh("testuser", "test-repo", tmp_path)
+        result = git_setup_module._create_repo_with_gh("testuser", "test-repo", tmp_path)
         assert result is False
 
 
@@ -100,7 +109,7 @@ def test_write_to_env_file_new_file(tmp_path, monkeypatch):
     
     # Change to tmp_path directory
     monkeypatch.chdir(tmp_path)
-    write_to_env_file("TEST_KEY", "test_value")
+    git_setup_module._write_to_env_file("TEST_KEY", "test_value")
     
     assert env_file.exists()
     content = env_file.read_text()
@@ -113,7 +122,7 @@ def test_write_to_env_file_update_existing(tmp_path, monkeypatch):
     env_file.write_text("EXISTING_KEY=old_value\n")
     
     monkeypatch.chdir(tmp_path)
-    write_to_env_file("EXISTING_KEY", "new_value")
+    git_setup_module._write_to_env_file("EXISTING_KEY", "new_value")
     content = env_file.read_text()
     assert "EXISTING_KEY=new_value" in content
     assert "old_value" not in content
@@ -125,80 +134,84 @@ def test_write_to_env_file_append_new_key(tmp_path, monkeypatch):
     env_file.write_text("EXISTING_KEY=value\n")
     
     monkeypatch.chdir(tmp_path)
-    write_to_env_file("NEW_KEY", "new_value")
+    git_setup_module._write_to_env_file("NEW_KEY", "new_value")
     content = env_file.read_text()
     assert "EXISTING_KEY=value" in content
     assert "NEW_KEY=new_value" in content
 
 
-def test_git_setup_command_existing_repo(tmp_path, monkeypatch):
+def test_git_setup_existing_repo(tmp_path, monkeypatch, mock_context):
     """Test git-setup when repo already exists."""
     # Create existing git repo
     calendar_dir = tmp_path / "calendars"
     calendar_dir.mkdir()
-    subprocess.run(["git", "init"], cwd=calendar_dir, check=True)
+    subprocess.run(["git", "init"], cwd=calendar_dir, check=True, capture_output=True)
     subprocess.run(
         ["git", "remote", "add", "origin", "https://github.com/user/repo.git"],
         cwd=calendar_dir,
         check=True,
+        capture_output=True,
     )
 
     config = CalendarConfig()
     config.calendar_dir = calendar_dir
+    mock_context._config = config
 
-    with patch("cli.commands.git_setup.CalendarConfig.from_env", return_value=config):
+    with patch.object(git_setup_module, "get_context", return_value=mock_context):
         # Should detect existing repo and remote
-        with patch("builtins.print") as mock_print:
-            git_setup_command()
+        with patch("typer.echo") as mock_echo:
+            git_setup_func()
             # Should print that repo already exists
-            assert any("already exists" in str(call) for call in mock_print.call_args_list)
+            assert any("already exists" in str(call) for call in mock_echo.call_args_list)
 
 
-def test_git_setup_command_init_new_repo(tmp_path, monkeypatch):
+def test_git_setup_init_new_repo(tmp_path, monkeypatch, mock_context):
     """Test git-setup initializing new repo."""
     calendar_dir = tmp_path / "calendars"
     calendar_dir.mkdir()
 
     config = CalendarConfig()
     config.calendar_dir = calendar_dir
+    mock_context._config = config
 
     original_cwd = os.getcwd()
     try:
         os.chdir(tmp_path)
-        with patch("cli.commands.git_setup.CalendarConfig.from_env", return_value=config), \
-             patch("cli.commands.git_setup.check_gh_cli_available", return_value=False), \
-             patch("builtins.input", return_value=""):
-            git_setup_command()
+        with patch.object(git_setup_module, "get_context", return_value=mock_context), \
+             patch.object(git_setup_module, "_check_gh_cli_available", return_value=False), \
+             patch("typer.prompt", return_value=""):
+            git_setup_func()
             # Should have created .git directory
             assert (calendar_dir / ".git").exists()
     finally:
         os.chdir(original_cwd)
 
 
-def test_git_setup_command_with_gh_cli(tmp_path, monkeypatch):
+def test_git_setup_with_gh_cli(tmp_path, monkeypatch, mock_context):
     """Test git-setup using GitHub CLI."""
     calendar_dir = tmp_path / "calendars"
     calendar_dir.mkdir()
 
     config = CalendarConfig()
     config.calendar_dir = calendar_dir
+    mock_context._config = config
 
     original_cwd = os.getcwd()
     try:
         os.chdir(tmp_path)
-        with patch("cli.commands.git_setup.CalendarConfig.from_env", return_value=config), \
-             patch("cli.commands.git_setup.check_gh_cli_available", return_value=True), \
-             patch("cli.commands.git_setup.get_github_username_from_gh", return_value="testuser"), \
-             patch("cli.commands.git_setup.create_repo_with_gh", return_value=True), \
-             patch("builtins.input", return_value=""):
-            git_setup_command()
+        with patch.object(git_setup_module, "get_context", return_value=mock_context), \
+             patch.object(git_setup_module, "_check_gh_cli_available", return_value=True), \
+             patch.object(git_setup_module, "_get_github_username_from_gh", return_value="testuser"), \
+             patch.object(git_setup_module, "_create_repo_with_gh", return_value=True), \
+             patch("typer.prompt", return_value=""):
+            git_setup_func()
             # Should have created .git directory
             assert (calendar_dir / ".git").exists()
     finally:
         os.chdir(original_cwd)
 
 
-def test_git_setup_command_creates_initial_commit(tmp_path, monkeypatch):
+def test_git_setup_creates_initial_commit(tmp_path, monkeypatch, mock_context):
     """Test git-setup creates initial commit if files exist."""
     calendar_dir = tmp_path / "calendars"
     calendar_dir.mkdir()
@@ -207,14 +220,15 @@ def test_git_setup_command_creates_initial_commit(tmp_path, monkeypatch):
 
     config = CalendarConfig()
     config.calendar_dir = calendar_dir
+    mock_context._config = config
 
     original_cwd = os.getcwd()
     try:
         os.chdir(tmp_path)
-        with patch("cli.commands.git_setup.CalendarConfig.from_env", return_value=config), \
-             patch("cli.commands.git_setup.check_gh_cli_available", return_value=False), \
-             patch("builtins.input", return_value=""):
-            git_setup_command()
+        with patch.object(git_setup_module, "get_context", return_value=mock_context), \
+             patch.object(git_setup_module, "_check_gh_cli_available", return_value=False), \
+             patch("typer.prompt", return_value=""):
+            git_setup_func()
             # Should have created .git directory
             assert (calendar_dir / ".git").exists()
             # Check if initial commit was created
