@@ -1,4 +1,4 @@
-"""Regression tests to ensure template processing matches legacy behavior."""
+"""Tests for template-based processing."""
 
 import tempfile
 from datetime import datetime
@@ -12,7 +12,7 @@ from app.ingestion.word_reader import WordReader
 from app.models.metadata import CalendarMetadata, CalendarWithMetadata
 from app.models.template_loader import load_template
 from app.output.ics_writer import ICSWriter
-from app.processing.event_processor import process_events, process_events_with_template
+from app.processing.event_processor import process_events_with_template
 
 
 def normalize_event_for_comparison(event):
@@ -27,56 +27,6 @@ def normalize_event_for_comparison(event):
     }
 
 
-def count_by_type(events):
-    """Count events by type."""
-    from collections import defaultdict
-
-    counts = defaultdict(int)
-    for event in events:
-        type_value = event.type or event.get_type_enum().value
-        counts[type_value] += 1
-    return dict(counts)
-
-
-def test_template_matches_legacy_output():
-    """Verify template processing produces equivalent results to current behavior."""
-    fixture_path = Path(__file__).parent / "fixtures" / "example-calendar.docx"
-    if not fixture_path.exists():
-        pytest.skip("Fixture file not found")
-
-    # Load templates
-    config = CalendarConfig.from_env()
-    medical_template = load_template("medical_default", config.template_dir)
-
-    # Process with legacy pipeline (using default template, then legacy processor)
-    reader = WordReader()
-    legacy_calendar = reader.read(fixture_path, medical_template)
-    legacy_events, legacy_summary = process_events(legacy_calendar.events)
-
-    # Process with template pipeline (using same template for both reading and processing)
-    template_calendar = reader.read(fixture_path, medical_template)
-    template_events, template_summary = process_events_with_template(
-        template_calendar.events, medical_template
-    )
-
-    # Compare event counts
-    legacy_counts = count_by_type(legacy_events)
-    template_counts = count_by_type(template_events)
-
-    # The counts should be similar (allowing for some differences in consolidation)
-    # This is a basic sanity check - full equivalence would require more detailed comparison
-    assert len(legacy_events) > 0
-    assert len(template_events) > 0
-
-    # Both should have processed events
-    assert legacy_summary["input_total"] > 0
-    assert template_summary["input_total"] > 0
-
-    # Both should have output events
-    assert legacy_summary["output_total"] > 0
-    assert template_summary["output_total"] > 0
-
-
 def test_template_word_reader_type_assignment():
     """Verify WordReader assigns types when template is provided."""
     fixture_path = Path(__file__).parent / "fixtures" / "example-calendar.docx"
@@ -89,7 +39,8 @@ def test_template_word_reader_type_assignment():
 
     # Read with template
     reader = WordReader()
-    calendar = reader.read(fixture_path, template)
+    ingestion_result = reader.read(fixture_path, template)
+    calendar = ingestion_result.calendar
 
     # Check that some events have types assigned
     events_with_types = [e for e in calendar.events if e.type is not None]
@@ -117,7 +68,8 @@ def test_end_to_end_template_vs_expected_ics():
 
     # Read and process with template
     word_reader = WordReader()
-    source_calendar = word_reader.read(docx_path, template)
+    ingestion_result = word_reader.read(docx_path, template)
+    source_calendar = ingestion_result.calendar
     processed_events, _ = process_events_with_template(source_calendar.events, template)
     processed_calendar = source_calendar.model_copy(update={"events": processed_events})
 
@@ -138,8 +90,10 @@ def test_end_to_end_template_vs_expected_ics():
 
             # Read both ICS files
             ics_reader = ICSReader()
-            actual_calendar = ics_reader.read(tmp_path)
-            expected_calendar = ics_reader.read(expected_ics_path)
+            actual_result = ics_reader.read(tmp_path)
+            expected_result = ics_reader.read(expected_ics_path)
+            actual_calendar = actual_result.calendar
+            expected_calendar = expected_result.calendar
 
             # Normalize events for comparison
             actual_normalized = sorted(
@@ -165,14 +119,14 @@ def test_end_to_end_template_vs_expected_ics():
 
             # Compare events (allowing for some differences in consolidation/formatting)
             # We'll do a fuzzy match - check that key events are present
-            actual_titles_by_date = {}
+            actual_titles_by_date: dict[str, list[str]] = {}
             for event in actual_normalized:
                 date_key = event["date"]
                 if date_key not in actual_titles_by_date:
                     actual_titles_by_date[date_key] = []
                 actual_titles_by_date[date_key].append(event["title"])
 
-            expected_titles_by_date = {}
+            expected_titles_by_date: dict[str, list[str]] = {}
             for event in expected_normalized:
                 date_key = event["date"]
                 if date_key not in expected_titles_by_date:
@@ -184,9 +138,9 @@ def test_end_to_end_template_vs_expected_ics():
             actual_dates = set(actual_titles_by_date.keys())
             expected_dates = set(expected_titles_by_date.keys())
             overlap = actual_dates & expected_dates
-            assert len(overlap) >= len(expected_dates) * 0.9, (
-                f"Too many date mismatches: {len(overlap)}/{len(expected_dates)} dates match"
-            )
+            assert (
+                len(overlap) >= len(expected_dates) * 0.9
+            ), f"Too many date mismatches: {len(overlap)}/{len(expected_dates)} dates match"
 
             # Sample a few key dates to verify key events are present
             sample_dates = sorted(overlap)[:20]  # Check first 20 dates
